Adds profile subcommand to miqperf
This command allows for a various benchmarking/profiling (memory, time,
stackprof, etc.) to be introduced into an existing ruby script, without
having to hardcode it into the script/codebase itself.

This assumes there is single entry point into the script, and creates a
temp script to run that dynamically adds the necessary profiling
utilities (with the minimum requirements to do so), and runs that in a
spawned process.

See the option parser for the description of all of the available
options.