[WIP] Add ManageIQPerformance.profile to profile subcommand
Makes use of the ManageIQPerformance.profile code in the `profile`
subcommand, and displays stats gained from that.

Has to do a decent amount of re-organizing in the code to allow the
proper lips to load first prior to called ManageIQPerformance.profile,
so there was a bit of churn.

Also added a stackprof report summary for both --miq-prof and
--stackprof when no stackprof output is defined, and the ability to
preload the rails application outside of the profiling code.

TODO:

- Better formatter for sql (reuse reporter code in some way?)