Fix cache_file location for downloaded extra gems
When .cache_file was called for gemspecs that were downloaded using
GemBase64.find_gemspec_for, it still looked for the .gem cache file as
if it was downloaded to the default rubygems directory.  This is because
the base_dir was not be generated properly after it was reset, since
there was some underlying logic that generates if from Gem.dir unless
@loaded_from is set, it would continue to generate the wrong path.

nil'ing out `@base_dir` and setting @loaded_from fixes this, though, it
is worth nothing that `@loaded_from` is technically set to the `.gem`
file, instead of the `.gemspec` file like it should, but it should be
fine.